Did quite some issue handling for the runtime loading of JARs.
The instrumentation agent is quite tricky to use because it relies on
the native "attach" library. The issue is that the library may only
be loaded once. Somehow (I did not quite get to the bottom of the issue)
when running a program using the agent with a minimal classpath
(i.e. especially without the tools.jar of the JDK), errors occur
relating to the already loaded native library when a first JAR
was already successfully loaded and now a second one should be loaded.
This does not happen when tools.jar is on the classpath. This seems
to have something to do with AttachProvider that is loaded by the
tools. Perhaps this issue will disappear with Java 9, but this is
just a hunch, I couldn't test it yet!
Thus, there are now to different strategies for Java up to version 8
and Java above version 8. Up to 8 we use another strategy that
involves using reflection to access the system class loader.
From java 9 on, this doesn't work any more and we use the agent
strategy. Note that using the agent requires the reference to the JAR
file including the agent. This JAR file also must have a manifest entry
with the "Agent-Class" key, pointing to the agent class.