Adding the PrerequisiteChecker.
The checker consists of a few simple classes that can be used to
check for references for being null and collections for being empty.
There is also support for Suppliers returning arbitrary objects that
can be checked for null or specifically collections that can be checked
for being empty. The main advantage of these classes is that
NullPointerExceptions happening internally in suppliers are reported
as such, making it possible to let null-chaining fail gracefully. Also,
all objects can be given names so one does exactly know what failed.
And finally, first all tests are performed and only all error messages
collected and then an exception is raised. This way, all errors are known
at once instead of restarting the test or application after each fix
only to run into the next exception.